+++ src/main/java/com/intellimarket/seller/dao/SellerDAO.java
@@ -1,0 +1,5 @@
+package com.intellimarket.seller.dao;
+
+public class SellerDAO {
+
+}
