+++ src/main/java/com/intellimarket/seller/exception/SellerException.java
@@ -1,0 +1,5 @@
+package com.intellimarket.seller.exception;
+
+public class SellerException {
+
+}
