--- src/main/java/com/intellimarket/storeadmin/dao/ProductImageDAO.java
@@ -1,5 +1,0 @@
-package com.intellimarket.storeadmin.dao;
-
-public class ProductImageDAO {
-
-}
