--- src/main/java/com/intellimarket/storeadmin/dao/SubCategoryDAO.java
@@ -1,5 +1,0 @@
-package com.intellimarket.storeadmin.dao;
-
-public class SubCategoryDAO {
-
-}
