--- src/main/java/com/intellimarket/storeadmin/exception/SellerException.java
@@ -1,5 +1,0 @@
-package com.intellimarket.storeadmin.exception;
-
-public class SellerException {
-
-}
