+++ src/main/resources/mapper/seller/ProductMapper.xml
@@ -1,3 +1,3 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.intellimarket.storeadmin.dao.ProductDAO"></mapper>
+<mapper namespace="com.intellimarket.seller.dao.ProductDAO"></mapper>
